AppDetails obtained and sorted/deserialized
The app can now pull all the information for apps, this allows the storage/sorting of them in the database. The current main issue is that in order to do this the program must run for ~4 days due to an api rate limit (200 requests per 5 minutes). See comments in GameService.java for more details on potential solutions.